normed layer
