+++ lib/iron_goat/lib/iron_goat/lib/src/layer.c
@@ -20,111 +20,6 @@ void destroy_iron_goat_layer(struct iron_goat_layer *self)
         self->properties->clear(&self->properties);
 }
 
-struct iron_goat_layer_type_parser {
-    ig_layer_type_t type;
-    char *match;
-};
-
-static const struct iron_goat_layer_type_parser LAYER_TOKEN_TYPE_PARSER[] = {
-    {LAYER_TILELAYER, "tilelayer"},
-    {LAYER_OBJECTGROUP, "objectgroup"},
-    {LAYER_IMAGELAYER, "imagelayer"},
-    {LAYER_GROUP, "group"}
-};
-
-bool load_iron_goat_layer_type(struct json *conf, struct iron_goat_layer *self)
-{
-    for (size_t i = 0; i < ARRAY_SIZE(LAYER_TOKEN_TYPE_PARSER); i++) {
-        if (estrcmp(conf->v.string, LAYER_TOKEN_TYPE_PARSER[i].match) == 0) {
-            self->type = LAYER_TOKEN_TYPE_PARSER[i].type;
-            return (true);
-        }
-    }
-    ASSERT("Json", "Layer type is unrecognized");
-    return (false);
-}
-
-bool load_iron_goat_layer_chunks(struct json *conf,
-    struct iron_goat_layer *self)
-{
-    struct json_array *data = conf->v.array;
-    struct iron_goat_chunk chnk = {0};
-
-    if ((self->chunks = VECTOR_CREATE(ig_chunk)) == NULL)
-        return (false);
-    for (size_t i = 0; i < data->size; i++) {
-        chnk = (struct iron_goat_chunk){0};
-        if (init_iron_goat_chunk(&data->data[i], &chnk) == false)
-            return (false);
-        if (self->chunks->push_back(&self->chunks, chnk) == -1)
-            return (false);
-    }
-    return (true);
-}
-
-bool load_iron_goat_layer_data(struct json *conf,
-    struct iron_goat_layer *self)
-{
-    struct json_array *data = conf->v.array;
-    u64_t val = {0};
-
-    if ((self->data = VECTOR_CREATE(ig_u64)) == NULL)
-        return (false);
-    for (size_t i = 0; i < data->size; i++) {
-        val = data->data[i].v.number;
-        if (self->data->push_back(&self->data, val) == -1)
-            return (false);
-    }
-    return (true);
-}
-
-struct iron_goat_layer_encoding_parser {
-    ig_layer_encoding_t type;
-    char *match;
-};
-
-static const struct iron_goat_layer_encoding_parser IG_ENCOD_TKNS_PARSER[] = {
-    {LAYER_CSV, "csv"},
-    {LAYER_BASE64, "base64"}
-};
-
-bool load_iron_goat_layer_encoding(struct json *conf,
-    struct iron_goat_layer *self)
-{
-    for (size_t i = 0; i < ARRAY_SIZE(IG_ENCOD_TKNS_PARSER); i++) {
-        if (estrcmp(conf->v.string, IG_ENCOD_TKNS_PARSER[i].match) == 0) {
-            self->encoding = IG_ENCOD_TKNS_PARSER[i].type;
-            return (true);
-        }
-    }
-    ASSERT("Json", "Layer encoding data is corrupted");
-    return (false);
-}
-
-
-struct iron_goat_layer_draworder_parser {
-    ig_layer_draworder_t type;
-    char *match;
-};
-
-static const struct iron_goat_layer_draworder_parser IG_DRORD_TKNS_PARSER[] = {
-    {LAYER_TOPDOWN, "topdown"},
-    {LAYER_INDEX,     "index"}
-};
-
-bool load_iron_goat_layer_draworder(struct json *conf,
-    struct iron_goat_layer *self)
-{
-    for (size_t i = 0; i < ARRAY_SIZE(IG_DRORD_TKNS_PARSER); i++) {
-        if (estrcmp(conf->v.string, IG_DRORD_TKNS_PARSER[i].match) == 0) {
-            self->draworder = IG_DRORD_TKNS_PARSER[i].type;
-            return (true);
-        }
-    }
-    ASSERT("Json", "Layer draworder data is corrupted");
-    return (false);
-}
-
 bool load_iron_goat_layer_objects(struct json *conf,
     struct iron_goat_layer *self)
 {
